+++ Natasha/Cards/PingCard.cs
@@ -44,12 +44,12 @@ internal sealed class PingCard : Card, IRegisterable, IHasCustomCardTraits
 				new ADrawCard { count = 2 },
 			],
 			Upgrade.A => [
-				new SequenceAction { CardId = uuid, SequenceStep = 1, SequenceLength = 2, Action = new AAttack { damage = GetDmg(s, 3) } },
-				new SequenceAction { CardId = uuid, SequenceStep = 2, SequenceLength = 2, Action = new ADrawCard { count = 3 } },
+				ModEntry.Instance.KokoroApi.Actions.MakeSequenceAction(uuid, 1, 2, new AAttack { damage = GetDmg(s, 3) }),
+				ModEntry.Instance.KokoroApi.Actions.MakeSequenceAction(uuid, 2, 2, new ADrawCard { count = 3 }),
 			],
 			_ => [
-				new SequenceAction { CardId = uuid, SequenceStep = 1, SequenceLength = 2, Action = new AAttack { damage = GetDmg(s, 2) } },
-				new SequenceAction { CardId = uuid, SequenceStep = 2, SequenceLength = 2, Action = new ADrawCard { count = 2 } },
+				ModEntry.Instance.KokoroApi.Actions.MakeSequenceAction(uuid, 1, 2, new AAttack { damage = GetDmg(s, 2) }),
+				ModEntry.Instance.KokoroApi.Actions.MakeSequenceAction(uuid, 2, 2, new ADrawCard { count = 2 }),
 			]
 		};
 }