+++ Natasha/ExternalAPIs/Kokoro/Actions/Sequence.cs
@@ -1,0 +1,9 @@
+﻿namespace Shockah.Natasha;
+
+public partial interface IKokoroApi
+{
+	public partial interface IActionApi
+	{
+		CardAction MakeSequenceAction(int cardId, int sequenceStep, int sequenceLength, CardAction action);
+	}
+}